Fix 2FA migration - user_id column type mismatch

+++ database/migrations/2026_07_13_000002_create_two_factor_sessions_table.php
@@ -10,7 +10,7 @@ class CreateTwoFactorSessionsTable extends Migration
     {
         Schema::create('two_factor_sessions', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('user_id');
+            $table->unsignedInteger('user_id');
             $table->string('token', 64)->unique();
             $table->boolean('verified')->default(false);
             $table->timestamp('expires_at');
